This commit was manufactured by cvs2svn to create branch 'THREADS_20060307_BRANCH'.

+++ mozilla/calendar/lightning/content/agenda-tree.js
@@ -149,10 +149,7 @@ function getCellText(row, column)
     if (event instanceof Synthetic)
         return "";
     var start = event.startDate || event.dueDate;
-    var dateFormatter = Components.classes["@mozilla.org/calendar/datetime-formatter;1"]
-                                  .getService(Components.interfaces.calIDateTimeFormatter);
-    start = start.getInTimezone(calendarDefaultTimezone());
-    return dateFormatter.formatDateTime(start);
+    return start.toString();
 };
 
 agendaTreeView.getLevel =
@@ -433,6 +430,9 @@ function observer_onAddItem(item)
 agendaTreeView.calendarObserver.onDeleteItem =
 function observer_onDeleteItem(item, rebuildFlag)
 {
+    if (!(item instanceof Components.interfaces.calIEvent))
+        return;
+
     var occs = item.getOccurrencesBetween(this.agendaTreeView.today.start,
                                           this.agendaTreeView.soon.end, {});
     occs.forEach(this.agendaTreeView.deleteItem, this.agendaTreeView);
